% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_regions.R
\name{generate_test_regions2}
\alias{generate_test_regions2}
\title{Generate set 2 of testing regions.}
\usage{
generate_test_regions2()
}
\value{
Nothing, saves files in the inst/testdata folder
}
\description{
Generate set 2 of testing regions.
}
\examples{

# no run

}
