% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{get_offset}
\alias{get_offset}
\title{Get offset}
\usage{
get_offset(gr_list, k, genome, refSet)
}
\arguments{
\item{gr_list}{a list of GRanges objects}

\item{k}{Number of sites to expand the oligonucleotide, size = 2*k + 1}

\item{genome}{a genome object from class BSgenome}

\item{refSet}{A reference set of positions.}
}
\value{
A data frame with oligonucleotide counts to use as offset.
}
\description{
Obtain the nucleotides at risk of each region.
}
\details{
It inputs the regions from \code{process_region_interactions2}.
}
\examples{

# not run

}
