% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{glm_nb_wrapper}
\alias{glm_nb_wrapper}
\title{NB regression wrapper}
\usage{
glm_nb_wrapper(
  data,
  formula,
  control = glm.control(),
  ci_parameters = NULL,
  ci_method = c("profile", "z", "none"),
  alpha = 0.05,
  name = "Regression"
)
}
\arguments{
\item{data}{A dataframe with counts and factors}

\item{formula}{A formula (as string or formula class)}

\item{control}{a glm.control object}

\item{ci_parameters}{parameters to compute confidence intervals}

\item{ci_method}{method to compute confidence intervals}

\item{alpha}{signifocance level for confidence intervals}

\item{name}{name of the experiment (logging)}
}
\value{
A data frame with coefficients and confidence intervals
}
\description{
It performs the negative bionomial regression from the MASS package.
It also calculates the confindence intervals using multiple methods
}
\examples{

# not yet

}
