% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{gr_intesect}
\alias{gr_intesect}
\title{Intersect a list of GRanges}
\usage{
gr_intesect(gr_base_list, ignore.strand = FALSE)
}
\arguments{
\item{gr_base_list}{list with GRanges}

\item{ignore.strand}{If to ignore strand when intersecting regions.}
}
\value{
a single GR object with the intersected sites
}
\description{
Given a list of GRanges objects it returns the intersection of all of them.
Thus, the resulting sites represent sites that are shared in all the
GRanges instances in the list.
}
\examples{

# no run.

}
