% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{obtain_oligo_counts}
\alias{obtain_oligo_counts}
\title{Oligo Counts}
\usage{
obtain_oligo_counts(gr, genome, k, refSet, data.frame = FALSE)
}
\arguments{
\item{gr}{a GRanges object}

\item{genome}{a genome object from class BSgenome}

\item{k}{Number of sites to expand the oligonucleotide, size = 2*k + 1}

\item{refSet}{A reference set of positions.}

\item{data.frame}{boolean if to return a dataframe or a vector}
}
\value{
A dataframe (or a vector) with the oligonucleotides contined in the
selected region.
}
\description{
Computes the olgonucleotide counts from GRanges object.
}
\examples{

library(GenomicRanges)
gr = GRanges(seqnames = c("chr1"),
        ranges = IRanges(start = c(1e6,2e6),width = 100))
genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

obtain_oligo_counts(gr = gr,
                    genome = genome,
                    k = 1,
                    refSet = c("C","A"),
                    data.frame = TRUE)

}
