% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{process_region_interactions2}
\alias{process_region_interactions2}
\title{Intersect regions}
\usage{
process_region_interactions2(list_of_regions)
}
\arguments{
\item{list_of_regions}{The list of regions obtained from \code{read_input_regions}}
}
\value{
a list of intersected regions
}
\description{
It performs a combinatorial intersection of regions. Basically, it
combines features in the different channels by intersectling them
with all the available combinations
}
\examples{

# no run
}
