% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{read_input_regions}
\alias{read_input_regions}
\title{Read input regions}
\usage{
read_input_regions(ref_data)
}
\arguments{
\item{ref_data}{a dataframe with the design table}
}
\value{
A list of list with the regions of interest.
}
\description{
It needs a design table with the files used as regions of interest.
It then reads and structures the bed files accordingly.
}
\details{
Used in regionMut region subcommand.
}
\examples{

# no run

}
